requirements.txt
# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
pydantic==2.5.3
pandas==2.2.0

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# Async Support
aiohttp==3.9.3
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-playwright==0.4.4
faker==22.6.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
